- .gitignore 머지 충돌 해결
- push를 막는 대용량 로그 파일 제거 (boot.log, spy.log, logs/)
- .gitignore에 로그 파일 제외 규칙 추가
- feat/notification/haechang 브랜치의 모든 변경사항 머지

+++ src/main/java/com/example/onlyone/domain/wallet/entity/Wallet.java
@@ -36,15 +36,15 @@ public class Wallet extends BaseTimeEntity {
 //    private int balance;
 
     @Column(name = "posted_balance")
-    private int postedBalance;
+    private Long postedBalance;
 
     @Column(name = "pending_out")
-    private int pendingOut;
+    private Long pendingOut;
 
     @OneToMany(mappedBy = "wallet", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<WalletTransaction> walletTransactions = new ArrayList<>();
 
-    public void updateBalance(int balance) {
+    public void updateBalance(Long balance) {
         this.postedBalance = balance;
     }
 }
